simplify implementatation
