+++ netket/models/__init__.py
@@ -22,12 +22,13 @@ from .deepset import DeepSetRelDistance, DeepSetMLP
 from .ndm import NDM
 from .autoreg import AbstractARNN, ARNNDense, ARNNConv1D, ARNNConv2D as _ARNNConv2D
 from .fast_autoreg import (
+    FastARNNSequential,
     FastARNNDense,
     FastARNNConv1D,
     FastARNNConv2D as _FastARNNConv2D,
 )
 from .mlp import MLP
-from .rnn import LSTMNet, GRUNet1D
+from .rnn import RNN, LSTMNet, GRUNet1D
 from .fast_rnn import FastLSTMNet, FastGRUNet1D
 from .utils import update_GCNN_parity
