+++ netket/nn/rnn/layers.py
@@ -14,7 +14,6 @@
 
 from typing import Callable, Optional
 
-import numpy as np
 from flax import linen as nn
 from flax.linen.dtypes import promote_dtype
 
@@ -25,62 +24,9 @@ from jax.nn.initializers import orthogonal, zeros
 from netket.utils import HashableArray
 from netket.utils.types import Array, DType, NNInitFunc
 
+from .ordering import check_reorder_idx
+
 default_kernel_init = orthogonal()
-
-
-def _check_reorder_idx(
-    reorder_idx: Optional[HashableArray],
-    inv_reorder_idx: Optional[HashableArray],
-    prev_neighbors: Optional[HashableArray],
-):
-    if reorder_idx is None and inv_reorder_idx is None and prev_neighbors is None:
-        # There is a faster code path for 1D RNN
-        return
-
-    if reorder_idx is None or inv_reorder_idx is None or prev_neighbors is None:
-        raise ValueError(
-            "`reorder_idx`, `inv_reorder_idx`, and `prev_neighbors` must be "
-            "provided at the same time."
-        )
-
-    reorder_idx = np.asarray(reorder_idx)
-    inv_reorder_idx = np.asarray(inv_reorder_idx)
-    prev_neighbors = np.asarray(prev_neighbors)
-
-    if reorder_idx.ndim != 1:
-        raise ValueError("`reorder_idx` must be 1D.")
-    if inv_reorder_idx.ndim != 1:
-        raise ValueError("`inv_reorder_idx` must be 1D.")
-    if prev_neighbors.ndim != 2:
-        raise ValueError("`prev_neighbors` must be 2D.")
-
-    V = reorder_idx.size
-    if inv_reorder_idx.size != V:
-        raise ValueError(
-            "`reorder_idx` and `inv_reorder_idx` must have the same length."
-        )
-    if prev_neighbors.shape[0] != V:
-        raise ValueError(
-            "`reorder_idx` and `prev_neighbors` must have the same length."
-        )
-
-    idx = reorder_idx[inv_reorder_idx]
-    if not np.array_equal(idx, np.arange(V)):
-        raise ValueError("`inv_reorder_idx` is not the inverse of `reorder_idx`.")
-
-    for i in range(V):
-        n = prev_neighbors[i]
-        for j in n:
-            if j < -1 or j >= V:
-                raise ValueError(f"Invaild neighbor {j} of site {i}")
-
-        n = [j for j in n if j >= 0]
-        if len(set(n)) != len(n):
-            raise ValueError(f"Duplicate neighbor {n} of site {i}")
-
-        for j in n:
-            if reorder_idx[j] >= reorder_idx[i]:
-                raise ValueError(f"Site {j} is not a previous neighbor of site {i}")
 
 
 class RNNLayer(nn.Module):
@@ -107,7 +53,7 @@ class RNNLayer(nn.Module):
 
     def __post_init__(self):
         super().__post_init__()
-        _check_reorder_idx(self.reorder_idx, self.inv_reorder_idx, self.prev_neighbors)
+        check_reorder_idx(self.reorder_idx, self.inv_reorder_idx, self.prev_neighbors)
 
     def _dense_params(self, name: Optional[str], in_features: int, out_features: int):
         kernel = self.param(
